Add files via upload
In this file, I used some calculation formulas and conditional formatting as well. I like that I am getting used to the use of these formulas. This assignment is as part of the google analytics program and I thought it as wise to share my version of what was taught by the tutor.